@@ -1,18 +1,18 @@
-import {Route} from "./path";
+import {RoutePath} from "./path";
 
 test("testPathParsing", () => {
-    let parsed = new Route("/#/page/?param1=2&param2=&param3");
+    let parsed = new RoutePath("/#/page/?param1=2&param2=&param3");
     expect(parsed.path).toBe("page");
     expect(parsed.params['param1']).toBe("2");
     expect(parsed.params['param2']).toBe(null);
     expect(parsed.params['param3']).toBe(true);
 
-    let parsed2 = new Route("/#/?param1=2&param2=&param3");
+    let parsed2 = new RoutePath("/#/?param1=2&param2=&param3");
     expect(parsed2.path).toBe("")
 
-    let parsed3 = new Route("?param1=2&param2=&param3");
+    let parsed3 = new RoutePath("?param1=2&param2=&param3");
     expect(parsed3.path).toBe("")
 
-    let parsed4 = new Route("?");
+    let parsed4 = new RoutePath("?");
     expect(parsed4.path).toBe("")
 });
